Fixed 64-bit transfers in the applet.

+++ clients/gnome-trans-applet/conduit.c
@@ -28,13 +28,13 @@ struct transfer *findtransferbytag(struct conduit *conduit, char *tag)
     return(transfer);
 }
 
-void transfersetsize(struct transfer *transfer, int size)
+void transfersetsize(struct transfer *transfer, intmax_t size)
 {
     transfer->size = size;
     cb_trsize(transfer, transfer->conduit->udata);
 }
 
-void transfersetpos(struct transfer *transfer, int pos)
+void transfersetpos(struct transfer *transfer, intmax_t pos)
 {
     transfer->pos = pos;
     cb_trpos(transfer, transfer->conduit->udata);
